add to cart and cart detail

+++ database/migrations/2024_01_24_190854_create_cart_items_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->unsignedInteger('cart_id');
             $table->unsignedInteger('product_id');
             $table->unsignedInteger('category_id');
-            $table->unsignedInteger('product_price');
+            $table->float('product_price');
             $table->float('product_with_add_on_price');
             $table->timestamps();
         });
